+++ relay-common/src/constants.rs
@@ -1,9 +1,11 @@
 //! Constants shared with the C-ABI and Sentry.
 
+// FIXME: Workaround for https://github.com/GREsau/schemars/pull/65
+#![allow(clippy::field_reassign_with_default)]
+
 use std::fmt;
 use std::str::FromStr;
 
-use failure::Fail;
 #[cfg(feature = "jsonschema")]
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -31,9 +33,16 @@ pub enum EventType {
 }
 
 /// An error used when parsing `EventType`.
-#[derive(Debug, Fail)]
-#[fail(display = "invalid event type")]
+#[derive(Clone, Copy, Debug)]
 pub struct ParseEventTypeError;
+
+impl fmt::Display for ParseEventTypeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "invalid event type")
+    }
+}
+
+impl std::error::Error for ParseEventTypeError {}
 
 impl Default for EventType {
     fn default() -> Self {
@@ -242,9 +251,16 @@ pub enum SpanStatus {
 }
 
 /// Error parsing a `SpanStatus`.
-#[derive(Debug, Fail)]
-#[fail(display = "invalid span status")]
+#[derive(Clone, Copy, Debug)]
 pub struct ParseSpanStatusError;
+
+impl fmt::Display for ParseSpanStatusError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "invalid span status")
+    }
+}
+
+impl std::error::Error for ParseSpanStatusError {}
 
 impl FromStr for SpanStatus {
     type Err = ParseSpanStatusError;