+++ relay-server/src/actors/events.rs
@@ -1,6 +1,7 @@
 use std::cell::RefCell;
 use std::collections::BTreeMap;
 use std::rc::Rc;
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 
@@ -10,7 +11,7 @@ use failure::Fail;
 use futures::prelude::*;
 use serde_json::Value as SerdeValue;
 
-use relay_common::{clone, metric, LogError, ProjectId};
+use relay_common::{clone, metric, ProjectId};
 use relay_config::{Config, RelayMode};
 use relay_general::pii::{PiiAttachmentsProcessor, PiiProcessor};
 use relay_general::processor::{process_value, ProcessingState};
@@ -20,6 +21,7 @@ use relay_general::protocol::{
 };
 use relay_general::store::ClockDriftProcessor;
 use relay_general::types::{Annotated, Array, Object, ProcessingAction, Value};
+use relay_log::LogError;
 use relay_quotas::RateLimits;
 use relay_redis::RedisPool;
 
@@ -30,7 +32,7 @@ use crate::actors::project::{
 use crate::actors::project_cache::ProjectError;
 use crate::actors::upstream::{SendRequest, UpstreamRelay, UpstreamRequestError};
 use crate::envelope::{self, AttachmentType, ContentType, Envelope, Item, ItemType};
-use crate::http::RequestBuilder;
+use crate::http::{HttpError, RequestBuilder};
 use crate::metrics::{RelayCounters, RelayHistograms, RelaySets, RelayTimers};
 use crate::service::ServerError;
 use crate::utils::{self, ChunkedFormDataAggregator, FormDataIter, FutureExt};
@@ -40,9 +42,7 @@ use {
     crate::actors::store::{StoreEnvelope, StoreError, StoreForwarder},
     crate::service::ServerErrorKind,
     crate::utils::EnvelopeLimiter,
-    chrono::TimeZone,
     failure::ResultExt,
-    minidump::Minidump,
     relay_filter::FilterStatKey,
     relay_general::store::{GeoIpLookup, StoreConfig, StoreProcessor},
     relay_quotas::{DataCategory, RateLimitingError, RedisRateLimiter},
@@ -128,6 +128,9 @@ enum ProcessingError {
 
     #[fail(display = "envelope empty, transaction removed by sampling")]
     TransactionSampled,
+
+    #[fail(display = "envelope empty, event removed by sampling")]
+    EventSampled,
 }
 
 impl ProcessingError {
@@ -167,6 +170,7 @@ impl ProcessingError {
 
             // Dynamic sampling (not an error, just discarding messages that were removed by sampling)
             Self::TransactionSampled => Some(Outcome::Invalid(DiscardReason::TransactionSampled)),
+            Self::EventSampled => Some(Outcome::Invalid(DiscardReason::EventSampled)),
 
             // If we send to an upstream, we don't emit outcomes.
             Self::SendFailed(_) => None,
@@ -321,11 +325,11 @@ impl EventProcessor {
             let mut session = match SessionUpdate::parse(&payload) {
                 Ok(session) => session,
                 Err(error) => {
-                    return sentry::with_scope(
+                    return relay_log::with_scope(
                         |s| s.set_extra("session", String::from_utf8_lossy(&payload).into()),
                         || {
                             // Skip gracefully here to allow sending other sessions.
-                            log::error!("failed to store session: {}", LogError(&error));
+                            relay_log::error!("failed to store session: {}", LogError(&error));
                             false
                         },
                     );
@@ -333,31 +337,31 @@ impl EventProcessor {
             };
 
             if session.sequence == u64::max_value() {
-                log::trace!("skipping session due to sequence overflow");
+                relay_log::trace!("skipping session due to sequence overflow");
                 return false;
             }
 
             if clock_drift_processor.is_drifted() {
-                log::trace!("applying clock drift correction to session");
+                relay_log::trace!("applying clock drift correction to session");
                 clock_drift_processor.process_session(&mut session);
                 changed = true;
             }
 
             if session.timestamp < session.started {
-                log::trace!("fixing session timestamp to {}", session.timestamp);
+                relay_log::trace!("fixing session timestamp to {}", session.timestamp);
                 session.timestamp = session.started;
                 changed = true;
             }
 
             let max_age = SignedDuration::seconds(self.config.max_session_secs_in_past());
             if (received - session.started) > max_age || (received - session.timestamp) > max_age {
-                log::trace!("skipping session older than {} days", max_age.num_days());
+                relay_log::trace!("skipping session older than {} days", max_age.num_days());
                 return false;
             }
 
             let max_future = SignedDuration::seconds(self.config.max_secs_in_future());
             if (session.started - received) > max_age || (session.timestamp - received) > max_age {
-                log::trace!(
+                relay_log::trace!(
                     "skipping session more than {}s in the future",
                     max_future.num_seconds()
                 );
@@ -374,7 +378,10 @@ impl EventProcessor {
             if changed {
                 let json_string = match serde_json::to_string(&session) {
                     Ok(json) => json,
-                    Err(_) => return false,
+                    Err(err) => {
+                        relay_log::error!("failed to serialize session: {}", LogError(&err));
+                        return false;
+                    }
                 };
 
                 item.set_payload(ContentType::Json, json_string);
@@ -384,21 +391,34 @@ impl EventProcessor {
         });
     }
 
-    /// Validates all user report/feedback items in the envelope, if any.
+    /// Validates and normalizes all user report items in the envelope.
     ///
     /// User feedback items are removed from the envelope if they contain invalid JSON or if the
-    /// JSON violates the schema (basic type validation).
+    /// JSON violates the schema (basic type validation). Otherwise, their normalized representation
+    /// is written back into the item.
     fn process_user_reports(&self, state: &mut ProcessEnvelopeState) {
         state.envelope.retain_items(|item| {
             if item.ty() != ItemType::UserReport {
                 return true;
             };
 
-            if let Err(error) = serde_json::from_slice::<UserReport>(&item.payload()) {
-                log::error!("failed to store user report: {}", LogError(&error));
-                return false;
-            }
+            let report = match serde_json::from_slice::<UserReport>(&item.payload()) {
+                Ok(session) => session,
+                Err(error) => {
+                    relay_log::error!("failed to store user report: {}", LogError(&error));
+                    return false;
+                }
+            };
 
+            let json_string = match serde_json::to_string(&report) {
+                Ok(json) => json,
+                Err(err) => {
+                    relay_log::error!("failed to serialize user report: {}", LogError(&err));
+                    return false;
+                }
+            };
+
+            item.set_payload(ContentType::Json, json_string);
             true
         });
     }
@@ -494,13 +514,21 @@ impl EventProcessor {
             .map_err(ProcessingError::InvalidJson)?
             .ok_or(ProcessingError::InvalidSecurityType)?;
 
-        match report_type {
+        let apply_result = match report_type {
             SecurityReportType::Csp => Csp::apply_to_event(data, &mut event),
             SecurityReportType::ExpectCt => ExpectCt::apply_to_event(data, &mut event),
             SecurityReportType::ExpectStaple => ExpectStaple::apply_to_event(data, &mut event),
             SecurityReportType::Hpkp => Hpkp::apply_to_event(data, &mut event),
+        };
+
+        if let Err(json_error) = apply_result {
+            // logged at call site of extract_event
+            relay_log::configure_scope(|scope| {
+                scope.set_extra("payload", String::from_utf8_lossy(&data).into());
+            });
+
+            return Err(ProcessingError::InvalidSecurityReport(json_error));
         }
-        .map_err(ProcessingError::InvalidSecurityReport)?;
 
         if let Some(release) = item.get_header("sentry_release").and_then(Value::as_str) {
             event.release = Annotated::from(LenientString(release.to_owned()));
@@ -535,7 +563,7 @@ impl EventProcessor {
                 // the optional `sentry` field.
                 match serde_json::from_str(entry.value()) {
                     Ok(event) => utils::merge_values(target, event),
-                    Err(_) => log::debug!("invalid json event payload in sentry form field"),
+                    Err(_) => relay_log::debug!("invalid json event payload in sentry form field"),
                 }
             } else if let Some(index) = utils::get_sentry_chunk_index(entry.key(), "sentry__") {
                 // Electron SDK splits up long payloads into chunks starting at sentry__1 with an
@@ -557,7 +585,7 @@ impl EventProcessor {
         if !aggregator.is_empty() {
             match serde_json::from_str(&aggregator.join()) {
                 Ok(event) => utils::merge_values(target, event),
-                Err(_) => log::debug!("invalid json event payload in sentry__* form fields"),
+                Err(_) => relay_log::debug!("invalid json event payload in sentry__* form fields"),
             }
         }
     }
@@ -722,27 +750,27 @@ impl EventProcessor {
         }
 
         let (event, event_len) = if let Some(item) = event_item.or(security_item) {
-            log::trace!("processing json event");
+            relay_log::trace!("processing json event");
             metric!(timer(RelayTimers::EventProcessingDeserialize), {
                 // Event items can never include transactions, so retain the event type and let
                 // inference deal with this during store normalization.
                 self.event_from_json_payload(item, None)?
             })
         } else if let Some(item) = transaction_item {
-            log::trace!("processing json transaction");
+            relay_log::trace!("processing json transaction");
             metric!(timer(RelayTimers::EventProcessingDeserialize), {
                 // Transaction items can only contain transaction events. Force the event type to
                 // hint to normalization that we're dealing with a transaction now.
                 self.event_from_json_payload(item, Some(EventType::Transaction))?
             })
         } else if let Some(item) = raw_security_item {
-            log::trace!("processing security report");
+            relay_log::trace!("processing security report");
             self.event_from_security_report(item)?
         } else if attachment_item.is_some() || breadcrumbs1.is_some() || breadcrumbs2.is_some() {
-            log::trace!("extracting attached event data");
+            relay_log::trace!("extracting attached event data");
             Self::event_from_attachments(&self.config, attachment_item, breadcrumbs1, breadcrumbs2)?
         } else if let Some(item) = form_item {
-            log::trace!("extracting form data");
+            relay_log::trace!("extracting form data");
             let len = item.len();
 
             let mut value = SerdeValue::Object(Default::default());
@@ -751,7 +779,7 @@ impl EventProcessor {
 
             (event, len)
         } else {
-            log::trace!("no event in envelope");
+            relay_log::trace!("no event in envelope");
             (Annotated::empty(), 0)
         };
 
@@ -771,75 +799,6 @@ impl EventProcessor {
             .map_err(ProcessingError::InvalidUnrealReport)
     }
 
-    /// Writes a placeholder to indicate that this event has an associated minidump or an apple
-    /// crash report.
-    ///
-    /// This will indicate to the ingestion pipeline that this event will need to be processed. The
-    /// payload can be checked via `is_minidump_event`.
-    #[cfg(feature = "processing")]
-    fn write_native_placeholder(&self, event: &mut Event, is_minidump: bool) {
-        use relay_general::protocol::{Exception, JsonLenientString, Level, Mechanism};
-
-        // Events must be native platform.
-        let platform = event.platform.value_mut();
-        *platform = Some("native".to_string());
-
-        // Assume that this minidump is the result of a crash and assign the fatal
-        // level. Note that the use of `setdefault` here doesn't generally allow the
-        // user to override the minidump's level as processing will overwrite it
-        // later.
-        event.level.get_or_insert_with(|| Level::Fatal);
-
-        // Create a placeholder exception. This signals normalization that this is an
-        // error event and also serves as a placeholder if processing of the minidump
-        // fails.
-        let exceptions = event
-            .exceptions
-            .value_mut()
-            .get_or_insert_with(Values::default)
-            .values
-            .value_mut()
-            .get_or_insert_with(Vec::new);
-
-        exceptions.clear(); // clear previous errors if any
-
-        let (type_name, value, mechanism_type) = if is_minidump {
-            ("Minidump", "Invalid Minidump", "minidump")
-        } else {
-            (
-                "AppleCrashReport",
-                "Invalid Apple Crash Report",
-                "applecrashreport",
-            )
-        };
-
-        exceptions.push(Annotated::new(Exception {
-            ty: Annotated::new(type_name.to_string()),
-            value: Annotated::new(JsonLenientString(value.to_string())),
-            mechanism: Annotated::new(Mechanism {
-                ty: Annotated::from(mechanism_type.to_string()),
-                handled: Annotated::from(false),
-                synthetic: Annotated::from(true),
-                ..Mechanism::default()
-            }),
-            ..Exception::default()
-        }));
-    }
-
-    /// Extracts the timestamp from the minidump and uses it as the event timestamp.
-    #[cfg(feature = "processing")]
-    fn write_minidump_timestamp(&self, event: &mut Event, minidump_item: &Item) {
-        let minidump = match Minidump::read(minidump_item.payload()) {
-            Ok(minidump) => minidump,
-            Err(err) => {
-                log::debug!("Failed to parse minidump: {:?}", err);
-                return;
-            }
-        };
-        let timestamp = Utc.timestamp(minidump.header.time_date_stamp.into(), 0);
-        event.timestamp.set_value(Some(timestamp.into()));
-    }
-
     /// Adds processing placeholders for special attachments.
     ///
     /// If special attachments are present in the envelope, this adds placeholder payloads to the
@@ -847,7 +806,7 @@ impl EventProcessor {
     ///
     /// If the event payload was empty before, it is created.
     #[cfg(feature = "processing")]
-    fn create_placeholders(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
+    fn create_placeholders(&self, state: &mut ProcessEnvelopeState) {
         let envelope = &mut state.envelope;
 
         let minidump_attachment =
@@ -858,15 +817,12 @@ impl EventProcessor {
         if let Some(item) = minidump_attachment {
             let event = state.event.get_or_insert_with(Event::default);
             state.metrics.bytes_ingested_event_minidump = Annotated::new(item.len() as u64);
-            self.write_native_placeholder(event, true);
-            self.write_minidump_timestamp(event, item);
+            utils::process_minidump(event, &item.payload());
         } else if let Some(item) = apple_crash_report_attachment {
             let event = state.event.get_or_insert_with(Event::default);
             state.metrics.bytes_ingested_event_applecrashreport = Annotated::new(item.len() as u64);
-            self.write_native_placeholder(event, false);
+            utils::process_apple_crash_report(event, &item.payload());
         }
-
-        Ok(())
     }
 
     fn finalize_event(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
@@ -936,7 +892,7 @@ impl EventProcessor {
             .and_then(|k| Some(k.numeric_id?.to_string()));
 
         if key_id.is_none() {
-            log::error!(
+            relay_log::error!(
                 "project state for key {} is missing key id",
                 envelope.meta().public_key()
             );
@@ -1063,7 +1019,7 @@ impl EventProcessor {
     /// This only applies the new PII rules that explicitly select `ValueType::Binary` or one of the
     /// attachment types. When special attachments are detected, these are scrubbed with custom
     /// logic; otherwise the entire attachment is treated as a single binary blob.
-    fn scrub_attachments(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
+    fn scrub_attachments(&self, state: &mut ProcessEnvelopeState) {
         let envelope = &mut state.envelope;
         if let Some(ref config) = state.project_state.config.pii_config {
             let minidump = envelope
@@ -1092,7 +1048,7 @@ impl EventProcessor {
                             timer(RelayTimers::MinidumpScrubbing) = start.elapsed(),
                             status = "error"
                         );
-                        log::warn!("failed to scrub minidump: {}", LogError(&scrub_error));
+                        relay_log::warn!("failed to scrub minidump: {}", LogError(&scrub_error));
                         metric!(timer(RelayTimers::AttachmentScrubbing), {
                             processor.scrub_attachment(filename, &mut payload);
                         })
@@ -1107,8 +1063,6 @@ impl EventProcessor {
                 item.set_payload(content_type, payload);
             }
         }
-
-        Ok(())
     }
 
     fn serialize_event(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
@@ -1125,6 +1079,21 @@ impl EventProcessor {
         state.envelope.add_item(event_item);
 
         Ok(())
+    }
+
+    /// Run dynamic sampling rules to see if we keep the event or remove it.
+    fn sample_event(&self, state: &mut ProcessEnvelopeState) -> Result<(), ProcessingError> {
+        let event = match &state.event.0 {
+            None => return Ok(()), // can't process without an event
+            Some(event) => event,
+        };
+
+        let project_id = state.project_id;
+        match utils::should_keep_event(event, &state.project_state, project_id) {
+            Some(false) => Err(ProcessingError::EventSampled),
+            Some(true) => Ok(()),
+            None => Ok(()), // Not enough info to make a definite evaluation, keep the event
+        }
     }
 
     fn process_state(
@@ -1148,16 +1117,18 @@ impl EventProcessor {
             });
 
             self.extract_event(&mut state).map_err(|error| {
-                log::error!("failed to extract event: {}", LogError(&error));
+                relay_log::error!("failed to extract event: {}", LogError(&error));
                 error
             })?;
 
             if_processing!({
                 self.process_unreal(&mut state)?;
-                self.create_placeholders(&mut state)?;
+                self.create_placeholders(&mut state);
             });
 
             self.finalize_event(&mut state)?;
+
+            self.sample_event(&mut state)?;
 
             if_processing!({
                 self.store_process_event(&mut state)?;
@@ -1174,7 +1145,7 @@ impl EventProcessor {
             self.serialize_event(&mut state)?;
         }
 
-        self.scrub_attachments(&mut state)?;
+        self.scrub_attachments(&mut state);
 
         Ok(ProcessEnvelopeResponse::from(state))
     }
@@ -1187,12 +1158,16 @@ impl EventProcessor {
 
         let project_id = state.project_id;
         let client = state.envelope.meta().client().map(str::to_owned);
+        let user_agent = state.envelope.meta().user_agent().map(str::to_owned);
 
-        sentry::with_scope(
+        relay_log::with_scope(
             |scope| {
                 scope.set_tag("project", project_id);
                 if let Some(client) = client {
                     scope.set_tag("sdk", client);
+                }
+                if let Some(user_agent) = user_agent {
+                    scope.set_extra("user_agent", user_agent.into());
                 }
             },
             || self.process_state(state),
@@ -1263,7 +1238,7 @@ impl EventManager {
         redis_pool: Option<RedisPool>,
     ) -> Result<Self, ServerError> {
         let thread_count = config.cpu_concurrency();
-        log::info!("starting {} event processing workers", thread_count);
+        relay_log::info!("starting {} event processing workers", thread_count);
 
         #[cfg(not(feature = "processing"))]
         let _ = redis_pool;
@@ -1325,11 +1300,11 @@ impl Actor for EventManager {
         // should ensure that we're not dropping events unintentionally after we've accepted them.
         let mailbox_size = self.config.event_buffer_size() as usize;
         context.set_mailbox_capacity(mailbox_size);
-        log::info!("event manager started");
+        relay_log::info!("event manager started");
     }
 
     fn stopped(&mut self, _ctx: &mut Self::Context) {
-        log::info!("event manager stopped");
+        relay_log::info!("event manager stopped");
     }
 }
 
@@ -1392,7 +1367,7 @@ impl Handler<QueueEnvelope> for EventManager {
         // that future will be tied to the EventManager's context. This allows to keep the Project
         // actor alive even if it is cleaned up in the ProjectManager.
 
-        log::trace!("queued event");
+        relay_log::trace!("queued event");
         Ok(event_id)
     }
 }
@@ -1453,6 +1428,7 @@ impl Handler<HandleEnvelope> for EventManager {
         let is_event = envelope.items().any(Item::creates_event);
 
         let scoping = Rc::new(RefCell::new(envelope.meta().get_partial_scoping()));
+        let is_received = Rc::new(AtomicBool::from(false));
 
         let future = project
             .send(CheckEnvelope::fetched(envelope))
@@ -1515,11 +1491,11 @@ impl Handler<HandleEnvelope> for EventManager {
                 }
             }))
             .into_actor(self)
-            .and_then(clone!(scoping, |mut envelope, slf, _| {
+            .and_then(clone!(scoping, is_received, |mut envelope, slf, _| {
                 #[cfg(feature = "processing")]
                 {
                     if let Some(store_forwarder) = store_forwarder {
-                        log::trace!("sending envelope to kafka");
+                        relay_log::trace!("sending envelope to kafka");
                         let future = store_forwarder
                             .send(StoreEnvelope {
                                 envelope,
@@ -1540,15 +1516,15 @@ impl Handler<HandleEnvelope> for EventManager {
                     // XXX: this is wrong because captured_events does not take envelopes without
                     // event_id into account.
                     if let Some(event_id) = event_id {
-                        log::debug!("capturing envelope");
+                        relay_log::debug!("capturing envelope");
                         slf.captures.insert(event_id, Ok(envelope));
                     } else {
-                        log::debug!("dropping non event envelope");
+                        relay_log::debug!("dropping non event envelope");
                     }
                     return Box::new(fut::ok(())) as ResponseActFuture<_, _, _>;
                 }
 
-                log::trace!("sending event to sentry endpoint");
+                relay_log::trace!("sending event to sentry endpoint");
                 let project_id = scoping.borrow().project_id;
                 let request = SendRequest::post(format!("/api/{}/envelope/", project_id)).build(
                     move |mut builder: RequestBuilder| {
@@ -1579,11 +1555,16 @@ impl Handler<HandleEnvelope> for EventManager {
                             .body(
                                 envelope
                                     .to_vec()
+                                    // XXX: upstream actor should allow for custom error type,
+                                    // right now we are forced to shoehorn our envelope errors into
+                                    // UpstreamRequestError
                                     .map_err(failure::Error::from)
-                                    .map_err(actix_web::Error::from)?
+                                    .map_err(actix_web::Error::from)
+                                    .map_err(HttpError::Actix)
+                                    .map_err(UpstreamRequestError::Http)?
                                     .into(),
                             )
-                            .map_err(UpstreamRequestError::from)
+                            .map_err(UpstreamRequestError::Http)
                     },
                 );
 
@@ -1591,6 +1572,15 @@ impl Handler<HandleEnvelope> for EventManager {
                     .send(request)
                     .map_err(ProcessingError::ScheduleFailed)
                     .and_then(move |result| {
+                        let received = match result {
+                            Ok(_) => true,
+                            Err(ref e) => e.is_received(),
+                        };
+
+                        // Flag that upstream has received the request, which will skip outcome
+                        // generation below.
+                        is_received.store(received, Ordering::SeqCst);
+
                         result.map_err(move |error| match error {
                             UpstreamRequestError::RateLimited(upstream_limits) => {
                                 let limits = upstream_limits.scope(&scoping.borrow());
@@ -1614,11 +1604,11 @@ impl Handler<HandleEnvelope> for EventManager {
                 if capture {
                     // XXX: does not work with envelopes without event_id
                     if let Some(event_id) = event_id {
-                        log::debug!("capturing failed event {}", event_id);
+                        relay_log::debug!("capturing failed event {}", event_id);
                         let msg = LogError(&error).to_string();
                         slf.captures.insert(event_id, Err(msg));
                     } else {
-                        log::debug!("dropping failed envelope without event");
+                        relay_log::debug!("dropping failed envelope without event");
                     }
                 }
 
@@ -1633,9 +1623,16 @@ impl Handler<HandleEnvelope> for EventManager {
                     // Errors are only logged for what we consider an internal discard reason. These
                     // indicate errors in the infrastructure or implementation bugs. In other cases,
                     // we "expect" errors and log them as debug level.
-                    log::error!("error processing event: {}", LogError(&error));
+                    relay_log::error!("error processing event: {}", LogError(&error));
                 } else {
-                    log::debug!("dropped event: {}", LogError(&error));
+                    relay_log::debug!("dropped event: {}", LogError(&error));
+                }
+
+                // Do not emit outcomes for requests that have been accepted by the upstream. In
+                // such a case, the upstream assumes ownership and logs the outcome, instead. This
+                // is irrelevant to processing Relays, since they never send to the upstream.
+                if is_received.load(Ordering::SeqCst) {
+                    return;
                 }
 
                 if let Some(outcome) = outcome {